+++ weblate/settings_docker.py
@@ -1103,7 +1103,6 @@ WEBLATE_MACHINERY = [
     "weblate.machinery.modernmt.ModernMTTranslation",
     "weblate.machinery.mymemory.MyMemoryTranslation",
     "weblate.machinery.netease.NeteaseSightTranslation",
-    "weblate.machinery.tmserver.AmagamaTranslation",
     "weblate.machinery.tmserver.TMServerTranslation",
     "weblate.machinery.yandex.YandexTranslation",
     "weblate.machinery.yandexv2.YandexV2Translation",